requirements.txt
torch==1.9.1
pytorch_lightning==1.4.7
hydra-core==1.1.1
wandb==0.12.2
xgboost==1.4.2
scikit-learn==0.24.2
matplotlib==3.4.3
optuna==2.9.1
openpyxl==3.0.8
hydra-colorlog==1.1.0
rich==10.10.0
einops==0.3.2
hydra-optuna-sweeper==1.1.1
captum==0.4.0

# dali by
# pip install --extra-index-url https://developer.download.nvidia.com/compute/redist/ nvidia-dali-cuda100==1.3.0
# decord